+++ sdmx-dl-format-xml/src/test/java/sdmxdl/xml/stream/XMLStreamCompactDataCursorTest.java
@@ -28,7 +28,7 @@ import sdmxdl.samples.SdmxSource;
 import sdmxdl.tck.DataCursorAssert;
 import sdmxdl.util.parser.DefaultObsParser;
 import sdmxdl.util.parser.FreqFactory;
-import sdmxdl.util.parser.PeriodParsers;
+import sdmxdl.util.parser.TimeFormatParsers;
 
 import javax.xml.stream.XMLInputFactory;
 import javax.xml.stream.XMLStreamException;
@@ -37,6 +37,7 @@ import java.io.InputStream;
 
 import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
+import static sdmxdl.util.parser.TimeFormatParsers.FIRST_DAY_OF_YEAR;
 
 /**
  * @author Philippe Charles
@@ -50,7 +51,7 @@ public class XMLStreamCompactDataCursorTest {
 
         DataCursorAssert.assertCompliance(() -> {
             InputStream stream = xml.openStream();
-            ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx20(FreqFactory.TIME_FORMAT_CONCEPT), PeriodParsers::onStandardFreq, Parser.onDouble());
+            ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx20(FreqFactory.TIME_FORMAT_CONCEPT), freq -> TimeFormatParsers.getObservationalTimePeriod(FIRST_DAY_OF_YEAR), Parser.onDouble());
             try {
                 return new XMLStreamCompactDataCursor(xif.createXMLStreamReader(stream), stream, builder, obsParser, "TIME_PERIOD", "OBS_VALUE");
             } catch (XMLStreamException e) {
@@ -58,7 +59,7 @@ public class XMLStreamCompactDataCursorTest {
             }
         }, Key.ALL, DataFilter.FULL);
 
-        ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx20(FreqFactory.TIME_FORMAT_CONCEPT), PeriodParsers::onStandardFreq, Parser.onDouble());
+        ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx20(FreqFactory.TIME_FORMAT_CONCEPT), freq -> TimeFormatParsers.getObservationalTimePeriod(FIRST_DAY_OF_YEAR), Parser.onDouble());
         try (InputStream stream = xml.openStream();
              DataCursor o = new XMLStreamCompactDataCursor(xif.createXMLStreamReader(stream), stream, builder, obsParser, "TIME_PERIOD", "OBS_VALUE")) {
             int indexSeries = -1;
@@ -100,7 +101,7 @@ public class XMLStreamCompactDataCursorTest {
 
         DataCursorAssert.assertCompliance(() -> {
             InputStream stream = xml.openStream();
-            ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx21(0), PeriodParsers::onStandardFreq, Parser.onDouble());
+            ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx21(0), freq -> TimeFormatParsers.getObservationalTimePeriod(FIRST_DAY_OF_YEAR), Parser.onDouble());
             try {
                 return new XMLStreamCompactDataCursor(xif.createXMLStreamReader(stream), stream, builder, obsParser, "TIME_PERIOD", "OBS_VALUE");
             } catch (XMLStreamException e) {
@@ -108,7 +109,7 @@ public class XMLStreamCompactDataCursorTest {
             }
         }, Key.ALL, DataFilter.FULL);
 
-        ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx21(0), PeriodParsers::onStandardFreq, Parser.onDouble());
+        ObsParser obsParser = new DefaultObsParser(FreqFactory.sdmx21(0), freq -> TimeFormatParsers.getObservationalTimePeriod(FIRST_DAY_OF_YEAR), Parser.onDouble());
         try (InputStream stream = xml.openStream();
              DataCursor o = new XMLStreamCompactDataCursor(xif.createXMLStreamReader(stream), stream, builder, obsParser, "TIME_PERIOD", "OBS_VALUE")) {
             assertThat(o.nextSeries()).isTrue();