+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/ri/web/drivers/BbkDriver.java
@@ -31,7 +31,7 @@ import sdmxdl.ext.ObsParser;
 import sdmxdl.util.SdmxFix;
 import sdmxdl.util.parser.DefaultObsParser;
 import sdmxdl.util.parser.FreqFactory;
-import sdmxdl.util.parser.PeriodParsers;
+import sdmxdl.util.parser.TimeFormatParsers;
 import sdmxdl.util.web.SdmxRestDriverSupport;
 import sdmxdl.web.SdmxWebSource;
 import sdmxdl.web.spi.SdmxWebContext;
@@ -42,6 +42,7 @@ import java.net.URL;
 import java.util.Objects;
 
 import static sdmxdl.util.SdmxFix.Category.QUERY;
+import static sdmxdl.util.parser.TimeFormatParsers.FIRST_DAY_OF_YEAR;
 
 /**
  * @author Philippe Charles
@@ -159,7 +160,7 @@ public final class BbkDriver implements SdmxWebDriver {
             Objects.requireNonNull(dsd);
             return new DefaultObsParser(
                     FreqFactory.sdmx20(dsd),
-                    PeriodParsers::onStandardFreq,
+                    freq -> TimeFormatParsers.getObservationalTimePeriod(FIRST_DAY_OF_YEAR),
                     Parser.onDouble()
             );
         }